Updated add product page to work without js

+++ app/controllers/customer/products_controller.rb
@@ -12,17 +12,17 @@ end
 def upload
 
   # check if file size is more of 2GB
-  if params[:product][:file] and (params[:product][:file].size < 2147483648) 
+  if params[:product][:file] and (params[:product][:file].size < 2147483648)
 
     name = request.uuid #+ '---' + params[:product][:file].original_filename
     directory = Rails.root.join('app', 'tmp_uploads').to_s
     path = File.join(directory, name)
-    File.open(path, "wb") { |f| 
+    File.open(path, "wb") { |f|
       f.write(params[:product][:file].read)
       f.chmod(0600)
     }
 
-    unless params[:product][:upload_id].blank? or (params[:product][:upload_id] == '1') 
+    unless params[:product][:upload_id].blank? or (params[:product][:upload_id] == '1')
       File.delete(File.join(directory, params[:product][:upload_id]))
     end
     msg = { status: 'ok', id: request.uuid}
@@ -37,21 +37,26 @@ def create
   params.permit(:product)
   @product = Product.new(params[:product].permit(:name, :price, :description, :thumb))
   @product.customer_id = current_customer.id
-  path = File.join(Rails.root.join('app', 'tmp_uploads').to_s, params[:product][:upload_id])
-  @product.file =  File.new(path, "r")
-  @product.file_file_name = params[:product][:upload_file_name]
-  
+  if params[:product][:file].present?
+    @product.file = params[:product][:file]
+  elsif params[:product][:upload_id].present?
+    path = File.join(Rails.root.join('app', 'tmp_uploads').to_s, params[:product][:upload_id])
+    @product.file =  File.new(path, "r")
+    @product.file_file_name = params[:product][:upload_file_name]
+  end
   if @product.save
    #generate slug
-   #@product.slug = 
+   #@product.slug =
    if @product.update(slug: Base52.encode(@product.id))
-   File.delete(path)
-   redirect_to customer_products_path, notice: 'Product was created.' 
+    unless params[:product][:upload_id].blank?
+     File.delete(path)
+   end
+   redirect_to customer_products_path, notice: 'Product was created.'
  end
- else
+else
   print @product.errors.messages
   render :nothing => true
-  end
+end
 end
 
 def show
@@ -80,9 +85,9 @@ def update
  if @product.update(params[:product].permit(:name, :price, :description, :thumb))
 
   redirect_to [:customer, @product]
-  else
-    render 'edit'
-  end
+else
+  render 'edit'
+end
 end
 def destroy
   @product = Product.find(params[:id])
@@ -91,6 +96,6 @@ def destroy
   end
   @product.destroy
   redirect_to customer_products_path
-  end
+end
 end
